+++ packages/rsnap-overlay/src/overlay.rs
@@ -9,7 +9,10 @@ mod cursor_runtime;
 mod hud_helpers;
 mod hud_runtime;
 mod image_helpers;
+mod input_runtime;
 mod rendering;
+mod scroll_capture_runtime;
+mod scroll_input_runtime;
 mod scroll_preview_runtime;
 mod scroll_runtime;
 mod session_state;
@@ -19,8 +22,6 @@ mod window_position_runtime;
 mod window_runtime;
 mod worker_runtime;
 
-#[cfg(target_os = "macos")]
-use std::collections::VecDeque;
 #[cfg(not(target_os = "macos"))]
 use std::env;
 #[cfg(target_os = "macos")]
@@ -1092,106 +1093,6 @@ impl OverlaySession {
 			self.request_redraw_toolbar_window();
 			self.schedule_egui_repaint_after(self.repaint_interval_for_monitor(self.state.monitor));
 		}
-	}
-
-	#[cfg(test)]
-	fn observe_scroll_capture_frame(
-		&mut self,
-		frame: RgbaImage,
-	) -> Option<Result<ScrollObserveOutcome>> {
-		self.observe_scroll_capture_frame_at(frame, Instant::now())
-	}
-
-	#[cfg(test)]
-	fn observe_scroll_capture_frame_at(
-		&mut self,
-		frame: RgbaImage,
-		observation_at: Instant,
-	) -> Option<Result<ScrollObserveOutcome>> {
-		self.observe_scroll_capture_frame_with_gate(frame, false, observation_at, false)
-	}
-
-	fn observe_scroll_capture_frame_with_gate(
-		&mut self,
-		frame: RgbaImage,
-		allow_stale_input: bool,
-		observation_at: Instant,
-		allow_post_stall_burst_search: bool,
-	) -> Option<Result<ScrollObserveOutcome>> {
-		let prior_block_reason = self.scroll_capture_observation_block_reason_at(observation_at);
-		#[cfg(target_os = "macos")]
-		let consumed_live_stream_stale_grace = !allow_stale_input
-			&& prior_block_reason == Some("stale_input")
-			&& self.consume_live_stream_stale_grace_if_current();
-		#[cfg(not(target_os = "macos"))]
-		let consumed_live_stream_stale_grace = false;
-		let allow_gate_bypass = allow_stale_input || consumed_live_stream_stale_grace;
-		let motion_rows_hint = self.scroll_capture_commit_motion_rows_hint_at(observation_at);
-
-		if !allow_gate_bypass && prior_block_reason.is_some() {
-			return Some(Ok(ScrollObserveOutcome::NoChange));
-		}
-
-		let result = {
-			let Some(session) = self.scroll_capture.session.as_mut() else {
-				self.scroll_capture_set_error("Scroll capture session is unavailable.");
-
-				return None;
-			};
-
-			session.observe_downward_sample_with_motion_hint_and_burst(
-				frame,
-				motion_rows_hint,
-				allow_post_stall_burst_search,
-			)
-		};
-
-		if let Ok(outcome) = &result {
-			self.consume_scroll_capture_downward_motion_rows_for_outcome(outcome);
-		}
-
-		Some(result)
-	}
-
-	fn scroll_capture_commit_motion_rows_hint_at(&self, observation_at: Instant) -> Option<u32> {
-		if self.scroll_capture.input_direction != Some(ScrollDirection::Down) {
-			return None;
-		}
-
-		let input_direction_at = self.scroll_capture.input_direction_at?;
-
-		if !self.scroll_capture.input_gesture_active
-			&& observation_at.saturating_duration_since(input_direction_at)
-				> SCROLL_CAPTURE_INPUT_FRESHNESS
-		{
-			return None;
-		}
-		if !self.scroll_capture.downward_motion_rows_pending.is_finite()
-			|| self.scroll_capture.downward_motion_rows_pending <= 0.0
-		{
-			return None;
-		}
-
-		Some(self.scroll_capture.downward_motion_rows_pending.ceil() as u32)
-	}
-
-	fn scroll_capture_set_error(&mut self, message: impl Into<String>) {
-		let message = message.into();
-
-		tracing::warn!(
-			op = "scroll_capture.error",
-			error = %message,
-			"Scroll capture paused on error."
-		);
-
-		if let Some(trace_recorder) = self.scroll_capture.trace_recorder.as_mut() {
-			trace_recorder.record_error(&message);
-		}
-
-		self.scroll_capture.paused = true;
-
-		self.state.set_error(message);
-		self.request_redraw_all();
 	}
 
 	fn pending_freeze_capture_matches(&self, monitor: MonitorRect) -> bool {
@@ -2967,17 +2868,19 @@ impl OverlaySession {
 
 		let should_trace_frozen_selection_drag_timing =
 			self.should_trace_frozen_selection_drag_timing();
-		let toolbar_position_elapsed = if should_trace_frozen_selection_drag_timing {
-			let toolbar_position_started_at = Instant::now();
-			let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
+		let toolbar_position_elapsed: Option<Duration> =
+			if should_trace_frozen_selection_drag_timing {
+				let toolbar_position_started_at = Instant::now();
+				let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
 
-			Some(toolbar_position_started_at.elapsed())
-		} else {
-			let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
+				Some(toolbar_position_started_at.elapsed())
+			} else {
+				let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
 
-			None
-		};
-		let redraw_request_elapsed = if should_trace_frozen_selection_drag_timing {
+				None
+			};
+		let redraw_request_elapsed: Option<Duration> = if should_trace_frozen_selection_drag_timing
+		{
 			let redraw_request_started_at = Instant::now();
 
 			self.request_redraw_for_monitor(monitor);
@@ -4661,357 +4564,6 @@ impl OverlaySession {
 		control
 	}
 
-	fn note_window_focus_change(&mut self, window_id: WindowId, focused: bool) {
-		if focused {
-			self.focused_window_ids.insert(window_id);
-
-			self.pending_focus_loss_cleanup = false;
-
-			return;
-		}
-
-		self.focused_window_ids.remove(&window_id);
-
-		if self.focused_window_ids.is_empty() {
-			self.pending_focus_loss_cleanup = true;
-		}
-	}
-
-	fn handle_toolbar_mouse_input(&mut self, state: ElementState) -> OverlayControl {
-		let toolbar_left_button_down = matches!(state, ElementState::Pressed);
-
-		if toolbar_left_button_down == self.toolbar_left_button_down {
-			return OverlayControl::Continue;
-		}
-		if toolbar_left_button_down {
-			self.toolbar_left_button_went_down = true;
-		} else {
-			self.toolbar_left_button_went_up = true;
-		}
-
-		self.toolbar_left_button_down = toolbar_left_button_down;
-
-		if !toolbar_left_button_down {
-			self.stop_frozen_selection_drag();
-			self.stop_frozen_mosaic_drag();
-
-			let _ = self.stop_frozen_text_edit_drag();
-
-			self.toolbar_state.dragging = false;
-			self.toolbar_state.drag_offset = Vec2::ZERO;
-			self.toolbar_state.drag_anchor = None;
-		} else {
-			self.toolbar_state.drag_offset = Vec2::ZERO;
-			self.toolbar_state.dragging = false;
-			self.toolbar_state.drag_anchor = None;
-		}
-
-		#[cfg(target_os = "macos")]
-		{
-			self.request_redraw_toolbar_window();
-
-			if !toolbar_left_button_down && self.frozen_text_edit.is_some() {
-				self.focus_frozen_text_input_window(self.state.monitor);
-			}
-		}
-
-		OverlayControl::Continue
-	}
-
-	fn reset_toolbar_pointer_state(&mut self) {
-		self.toolbar_left_button_down = false;
-		self.toolbar_left_button_went_down = false;
-		self.toolbar_left_button_went_up = false;
-		self.toolbar_pointer_local = None;
-		self.toolbar_state.drag_anchor = None;
-	}
-
-	fn handle_modifiers_changed(&mut self, modifiers: &Modifiers) -> OverlayControl {
-		self.keyboard_modifiers = modifiers.state();
-
-		OverlayControl::Continue
-	}
-
-	#[cfg(not(target_os = "macos"))]
-	fn sample_mouse_location(&mut self) -> GlobalPoint {
-		let Some(cursor_device) = self.cursor_device.as_ref() else {
-			return GlobalPoint::new(0, 0);
-		};
-		let mouse = cursor_device.get_mouse();
-
-		GlobalPoint::new(mouse.coords.0, mouse.coords.1)
-	}
-
-	#[cfg(target_os = "macos")]
-	fn sample_mouse_location(&mut self) -> GlobalPoint {
-		let started_at = Instant::now();
-		let point = macos_mouse_location().unwrap_or(GlobalPoint::new(0, 0));
-		let elapsed = started_at.elapsed();
-
-		self.slow_op_logger.warn_if_slow(
-			"overlay.macos_cursor_location",
-			elapsed,
-			SLOW_OP_WARN_CURSOR_LOCATION,
-			|| format!("sample point=({}, {})", point.x, point.y),
-		);
-
-		point
-	}
-
-	fn last_fresh_event_cursor(&self) -> Option<(MonitorRect, GlobalPoint)> {
-		self.last_fresh_event_cursor_with_ttl(CURSOR_EVENT_TICK_TTL)
-	}
-
-	fn last_fresh_event_cursor_with_ttl(
-		&self,
-		ttl: Duration,
-	) -> Option<(MonitorRect, GlobalPoint)> {
-		let event_cursor_at = self.last_event_cursor_at?;
-		let event_cursor = self.last_event_cursor?;
-
-		if event_cursor_at.elapsed() > ttl {
-			return None;
-		}
-
-		Some(event_cursor)
-	}
-
-	fn set_alt_held(&mut self, alt: bool) {
-		if self.state.alt_held == alt {
-			return;
-		}
-		if alt && !matches!(self.state.mode, OverlayMode::Live) {
-			return;
-		}
-
-		self.state.alt_held = alt;
-
-		if !alt {
-			self.handle_alt_release();
-
-			return;
-		}
-		if self.live_drag_hides_auxiliary_windows() {
-			self.state.loupe = None;
-
-			self.set_alt_loupe_window_visible(None, false);
-
-			return;
-		}
-
-		let Some((monitor, cursor)) = self.alt_activation_cursor_context() else {
-			return;
-		};
-
-		self.set_alt_loupe_window_visible(Some(monitor), true);
-
-		if self.use_fake_hud_blur() {
-			self.maybe_request_live_bg(monitor);
-		}
-
-		self.request_live_alt_samples(monitor, cursor);
-	}
-
-	fn apply_loupe_activation_input(&mut self, pressed: bool, repeat: bool) -> bool {
-		let previous_alt_held = self.state.alt_held;
-
-		match self.config.alt_activation {
-			AltActivationMode::Hold => self.set_alt_held(pressed),
-			AltActivationMode::Toggle => {
-				if pressed && !repeat {
-					self.set_alt_held(!self.state.alt_held);
-				}
-			},
-		}
-
-		previous_alt_held != self.state.alt_held
-	}
-
-	fn apply_loupe_activation_key_event(&mut self, pressed: bool, repeat: bool) -> bool {
-		if !matches!(self.state.mode, OverlayMode::Live) {
-			self.loupe_activation_key_down = false;
-
-			return false;
-		}
-
-		self.loupe_activation_key_down = pressed;
-
-		if !pressed && !self.state.alt_held {
-			return false;
-		}
-		if pressed && !self.loupe_activation_shortcut_available() {
-			return false;
-		}
-
-		self.apply_loupe_activation_input(pressed, repeat)
-	}
-
-	fn clear_loupe_activation_on_focus_loss(&mut self) {
-		let should_reset = self.loupe_activation_key_down
-			|| (matches!(self.config.alt_activation, AltActivationMode::Hold)
-				&& self.state.alt_held);
-
-		if !should_reset {
-			return;
-		}
-
-		self.loupe_activation_key_down = false;
-
-		if self.apply_loupe_activation_input(false, false) {
-			let _ = self.request_redraw_for_alt_state_change();
-		}
-	}
-
-	fn maybe_clear_loupe_activation_after_focus_loss(&mut self) {
-		if !self.pending_focus_loss_cleanup || !self.focused_window_ids.is_empty() {
-			return;
-		}
-
-		self.pending_focus_loss_cleanup = false;
-
-		self.clear_loupe_activation_on_focus_loss();
-	}
-
-	fn request_redraw_for_alt_state_change(&mut self) -> OverlayControl {
-		if matches!(self.state.mode, OverlayMode::Live) {
-			self.request_redraw_hud_window();
-
-			if !self.live_loupe_uses_hud_window()
-				&& (self.state.alt_held || self.loupe_window_visible)
-			{
-				self.request_redraw_loupe_window();
-			}
-		}
-
-		OverlayControl::Continue
-	}
-
-	fn alt_activation_cursor_context(&mut self) -> Option<(MonitorRect, GlobalPoint)> {
-		if let Some((monitor, cursor)) = self.last_fresh_event_cursor() {
-			self.seed_alt_activation_cursor_context(monitor, cursor);
-
-			return Some((monitor, cursor));
-		}
-
-		let cursor = self.sample_mouse_location();
-		let Some(monitor) = self.monitor_at(cursor) else {
-			if self.state.cursor.is_none() {
-				self.state.cursor = Some(cursor);
-			}
-
-			return self.active_cursor_monitor().zip(self.state.cursor);
-		};
-
-		self.seed_alt_activation_cursor_context(monitor, cursor);
-
-		Some((monitor, cursor))
-	}
-
-	fn seed_alt_activation_cursor_context(&mut self, monitor: MonitorRect, cursor: GlobalPoint) {
-		let old_monitor = self.active_cursor_monitor();
-		let old_cursor = self.state.cursor;
-
-		match self.state.mode {
-			OverlayMode::Live => {
-				self.update_cursor_for_live_move(old_monitor, old_cursor, monitor, cursor)
-			},
-			OverlayMode::Frozen => self.update_cursor_state(monitor, cursor),
-		}
-	}
-
-	fn handle_alt_release(&mut self) {
-		self.state.loupe = None;
-		self.loupe_outer_pos = None;
-		self.pending_loupe_outer_pos = None;
-
-		self.set_alt_loupe_window_visible(None, false);
-
-		if matches!(self.state.mode, OverlayMode::Live) {
-			self.request_redraw_hud_window();
-		}
-	}
-
-	fn set_alt_loupe_window_visible(&mut self, monitor: Option<MonitorRect>, visible: bool) {
-		if !matches!(self.state.mode, OverlayMode::Live) {
-			self.loupe_window_visible = false;
-
-			self.reset_loupe_window_warmup_redraws();
-
-			if let Some(loupe_window) = self.loupe_window.as_ref() {
-				loupe_window.window.set_visible(false);
-				loupe_window.window.request_redraw();
-			}
-
-			return;
-		}
-		if self.live_loupe_uses_hud_window() {
-			self.loupe_window_visible = false;
-
-			self.reset_loupe_window_warmup_redraws();
-
-			if let Some(loupe_window) = self.loupe_window.as_ref() {
-				loupe_window.window.set_visible(false);
-			}
-
-			return;
-		}
-		if visible {
-			let Some(monitor) = monitor else {
-				return;
-			};
-
-			#[cfg(target_os = "macos")]
-			if self.loupe_window.is_none() {
-				self.request_aux_window_creation_if_needed();
-
-				return;
-			}
-
-			self.maybe_apply_pending_startup_aux_live_stream_filter_upgrade(monitor);
-
-			let visible = self.update_loupe_window_position(monitor);
-			let was_visible = self.loupe_window_visible;
-
-			self.loupe_window_visible = visible;
-
-			if visible {
-				self.force_apply_pending_loupe_window_move();
-			}
-			if visible {
-				if !was_visible {
-					self.maybe_start_loupe_window_warmup_redraw();
-				}
-			} else {
-				self.reset_loupe_window_warmup_redraws();
-			}
-
-			if let Some(loupe_window) = self.loupe_window.as_ref() {
-				loupe_window.window.set_visible(visible);
-				loupe_window.window.request_redraw();
-			}
-
-			return;
-		}
-
-		self.loupe_window_visible = false;
-
-		self.reset_loupe_window_warmup_redraws();
-
-		if let Some(loupe_window) = self.loupe_window.as_ref() {
-			loupe_window.window.set_visible(false);
-			loupe_window.window.request_redraw();
-		}
-	}
-
-	fn request_live_alt_samples(&mut self, monitor: MonitorRect, cursor: GlobalPoint) {
-		let sample_updated = self.request_live_cursor_sample(monitor, cursor, true);
-		let apply = self.live_sample_request_redraw_intent(false, sample_updated, true);
-
-		if apply.any_changed() {
-			self.request_redraw_live_sample_targets(monitor, apply);
-		}
-	}
-
 	fn handle_resized(&mut self, window_id: WindowId, size: PhysicalSize<u32>) -> OverlayControl {
 		let window_scale_factor = self
 			.windows
@@ -5119,1221 +4671,6 @@ impl OverlaySession {
 		}
 	}
 
-	fn handle_cursor_moved(
-		&mut self,
-		window_id: WindowId,
-		position: PhysicalPosition<f64>,
-	) -> OverlayControl {
-		let old_monitor = self.active_cursor_monitor();
-		let Some(overlay_window) = self.windows.get(&window_id) else {
-			return self.handle_cursor_moved_without_overlay_window(window_id, old_monitor);
-		};
-		let window_monitor = overlay_window.monitor;
-		let scale_factor = overlay_window.window.scale_factor();
-		let window_size = overlay_window.window.inner_size();
-
-		self.handle_cursor_moved_with_overlay_window(
-			window_id,
-			position,
-			old_monitor,
-			window_monitor,
-			scale_factor,
-			window_size,
-		)
-	}
-
-	fn handle_cursor_moved_with_overlay_window(
-		&mut self,
-		window_id: WindowId,
-		position: PhysicalPosition<f64>,
-		old_monitor: Option<MonitorRect>,
-		window_monitor: MonitorRect,
-		scale_factor: f64,
-		window_size: PhysicalSize<u32>,
-	) -> OverlayControl {
-		let should_trace_frozen_selection_drag_timing =
-			self.should_trace_frozen_selection_drag_timing();
-		let cursor_move_started_at = should_trace_frozen_selection_drag_timing.then(Instant::now);
-		let now = Instant::now();
-		let event_global = Self::overlay_window_event_global_position(
-			window_monitor,
-			scale_factor,
-			window_size,
-			position,
-		);
-		let frozen_selection_drag_global = self.frozen_selection_drag_cursor_move_global(
-			window_monitor,
-			scale_factor,
-			window_size,
-			position,
-			event_global,
-		);
-		let monitor = window_monitor;
-		let global = event_global;
-		let source = DeviceCursorPointSource::EventRecentFallback;
-		let device_cursor = event_global;
-
-		self.last_event_cursor = Some((monitor, event_global));
-		self.last_event_cursor_at = Some(now);
-
-		let old_cursor = self.state.cursor;
-		let trace = CursorMoveTrace {
-			window_id,
-			position,
-			old_cursor,
-			device_cursor,
-			event_global,
-			monitor,
-			global,
-			source,
-		};
-
-		self.trace_cursor_moved_with_mapping(trace);
-
-		let timing = self.run_cursor_move_updates(
-			should_trace_frozen_selection_drag_timing,
-			cursor_move_started_at,
-			old_monitor,
-			monitor,
-			global,
-			frozen_selection_drag_global,
-		);
-
-		if should_trace_frozen_selection_drag_timing {
-			self.trace_frozen_selection_drag_cursor_move(monitor, old_monitor, old_cursor, timing);
-		}
-		if self.update_frozen_brush_stroke(global) {
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		OverlayControl::Continue
-	}
-
-	fn overlay_window_event_global_position(
-		window_monitor: MonitorRect,
-		scale_factor: f64,
-		window_size: PhysicalSize<u32>,
-		position: PhysicalPosition<f64>,
-	) -> GlobalPoint {
-		let scale_factor = scale_factor.max(f64::MIN_POSITIVE);
-		let logical_width = ((window_size.width as f64) / scale_factor).max(1.0);
-		let logical_height = ((window_size.height as f64) / scale_factor).max(1.0);
-		let max_local_x = logical_width as i32 - 1;
-		let max_local_y = logical_height as i32 - 1;
-		let local_x = (position.x / scale_factor).round() as i32;
-		let local_y = (position.y / scale_factor).round() as i32;
-
-		GlobalPoint::new(
-			window_monitor.origin.x + local_x.clamp(0, max_local_x),
-			window_monitor.origin.y + local_y.clamp(0, max_local_y),
-		)
-	}
-
-	fn frozen_selection_drag_cursor_move_global(
-		&self,
-		window_monitor: MonitorRect,
-		scale_factor: f64,
-		window_size: PhysicalSize<u32>,
-		position: PhysicalPosition<f64>,
-		default_global: GlobalPoint,
-	) -> GlobalPoint {
-		if !matches!(self.state.mode, OverlayMode::Frozen) || !self.frozen_selection_drag.active {
-			return default_global;
-		}
-
-		Self::overlay_window_frozen_selection_drag_global_position(
-			window_monitor,
-			scale_factor,
-			window_size,
-			position,
-		)
-	}
-
-	fn overlay_window_frozen_selection_drag_global_position(
-		window_monitor: MonitorRect,
-		scale_factor: f64,
-		window_size: PhysicalSize<u32>,
-		position: PhysicalPosition<f64>,
-	) -> GlobalPoint {
-		let scale_factor = scale_factor.max(f64::MIN_POSITIVE);
-		let logical_width = ((window_size.width as f64) / scale_factor).max(1.0);
-		let logical_height = ((window_size.height as f64) / scale_factor).max(1.0);
-		let max_local_x = logical_width.ceil() as i32 - 1;
-		let max_local_y = logical_height.ceil() as i32 - 1;
-		// Frozen selection dragging should treat the event as covering the current logical cell so
-		// the final fractional trackpad move cannot nudge the rect lower-right on release.
-		let local_x = ((position.x / scale_factor).floor() as i32).clamp(0, max_local_x);
-		let local_y = ((position.y / scale_factor).floor() as i32).clamp(0, max_local_y);
-
-		GlobalPoint::new(window_monitor.origin.x + local_x, window_monitor.origin.y + local_y)
-	}
-
-	fn handle_cursor_moved_without_overlay_window(
-		&mut self,
-		window_id: WindowId,
-		old_monitor: Option<MonitorRect>,
-	) -> OverlayControl {
-		let should_trace_frozen_selection_drag_timing =
-			self.should_trace_frozen_selection_drag_timing();
-		let cursor_move_started_at = should_trace_frozen_selection_drag_timing.then(Instant::now);
-
-		if self.should_ignore_live_auxiliary_cursor_event(window_id) {
-			return OverlayControl::Continue;
-		}
-
-		let now = Instant::now();
-		let raw = self.sample_mouse_location();
-		let Some((monitor, global, source)) = self.resolve_device_cursor_point(raw) else {
-			return OverlayControl::Continue;
-		};
-		let old_cursor = self.state.cursor;
-
-		self.last_event_cursor = Some((monitor, global));
-		self.last_event_cursor_at = Some(now);
-
-		if tracing::enabled!(tracing::Level::TRACE) {
-			tracing::trace!(
-				window_id = ?window_id,
-				window_known = false,
-				old_cursor = ?old_cursor,
-				device_cursor = ?global,
-				event_cursor = ?global,
-				source = source.as_str(),
-				"CursorMoved (no overlay window mapping)."
-			);
-		}
-
-		let timing = self.run_cursor_move_updates(
-			should_trace_frozen_selection_drag_timing,
-			cursor_move_started_at,
-			old_monitor,
-			monitor,
-			global,
-			global,
-		);
-
-		if should_trace_frozen_selection_drag_timing {
-			self.trace_frozen_selection_drag_cursor_move(monitor, old_monitor, old_cursor, timing);
-		}
-		if self.update_frozen_brush_stroke(global) {
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		OverlayControl::Continue
-	}
-
-	fn run_cursor_move_updates(
-		&mut self,
-		should_trace_frozen_selection_drag_timing: bool,
-		cursor_move_started_at: Option<Instant>,
-		old_monitor: Option<MonitorRect>,
-		monitor: MonitorRect,
-		global: GlobalPoint,
-		frozen_selection_drag_global: GlobalPoint,
-	) -> FrozenSelectionDragCursorMoveTiming {
-		let old_cursor = self.state.cursor;
-		let cursor_update_elapsed =
-			Self::measure_duration_if(should_trace_frozen_selection_drag_timing, || {
-				self.update_cursor_for_live_move(old_monitor, old_cursor, monitor, global)
-			});
-		let previous_drag_rect = self.state.drag_rect;
-		let live_drag_update_elapsed =
-			Self::measure_duration_if(should_trace_frozen_selection_drag_timing, || {
-				self.update_live_drag_rect(monitor, global);
-			});
-		let (frozen_rect_changed, frozen_drag_update_elapsed) =
-			if should_trace_frozen_selection_drag_timing {
-				let frozen_drag_update_started_at = Instant::now();
-				let frozen_rect_changed =
-					self.update_frozen_selection_drag_rect(frozen_selection_drag_global);
-
-				self.update_frozen_mosaic_drag_rect(global);
-				self.update_frozen_text_edit_drag_anchor(global);
-
-				(frozen_rect_changed, Some(frozen_drag_update_started_at.elapsed()))
-			} else {
-				let frozen_rect_changed =
-					self.update_frozen_selection_drag_rect(frozen_selection_drag_global);
-
-				self.update_frozen_mosaic_drag_rect(global);
-				self.update_frozen_text_edit_drag_anchor(global);
-
-				(frozen_rect_changed, None)
-			};
-		let sync_cursor_icons_elapsed =
-			Self::measure_duration_if(should_trace_frozen_selection_drag_timing, || {
-				self.sync_overlay_cursor_icons();
-			});
-		let request_samples_elapsed =
-			Self::measure_duration_if(should_trace_frozen_selection_drag_timing, || {
-				self.request_cursor_move_samples(monitor, global);
-			});
-
-		if let Some(old_monitor) = old_monitor
-			&& old_monitor != monitor
-		{
-			self.request_redraw_for_monitor(old_monitor);
-		}
-
-		if Self::live_overlay_redraw_needed_for_cursor_update(
-			old_monitor,
-			monitor,
-			previous_drag_rect,
-			self.state.drag_rect,
-		) {
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		FrozenSelectionDragCursorMoveTiming {
-			cursor_update_elapsed: cursor_update_elapsed.unwrap_or_default(),
-			live_drag_update_elapsed: live_drag_update_elapsed.unwrap_or_default(),
-			frozen_drag_update_elapsed: frozen_drag_update_elapsed.unwrap_or_default(),
-			frozen_rect_changed,
-			sync_cursor_icons_elapsed: sync_cursor_icons_elapsed.unwrap_or_default(),
-			request_samples_elapsed: request_samples_elapsed.unwrap_or_default(),
-			total_elapsed: cursor_move_started_at
-				.map_or(Duration::ZERO, |started_at| started_at.elapsed()),
-		}
-	}
-
-	fn measure_duration_if(enabled: bool, operation: impl FnOnce()) -> Option<Duration> {
-		if enabled {
-			let started_at = Instant::now();
-
-			operation();
-
-			Some(started_at.elapsed())
-		} else {
-			operation();
-
-			None
-		}
-	}
-
-	fn should_ignore_live_auxiliary_cursor_event(&self, window_id: WindowId) -> bool {
-		Self::should_ignore_live_auxiliary_cursor_event_for_role(
-			self.state.mode,
-			self.is_auxiliary_capture_window(window_id),
-		)
-	}
-
-	fn is_auxiliary_capture_window(&self, window_id: WindowId) -> bool {
-		self.hud_window.as_ref().is_some_and(|window| window.window.id() == window_id)
-			|| self.loupe_window.as_ref().is_some_and(|window| window.window.id() == window_id)
-			|| self.toolbar_window.as_ref().is_some_and(|window| window.window.id() == window_id)
-			|| self
-				.scroll_preview_window
-				.as_ref()
-				.is_some_and(|window| window.window.id() == window_id)
-	}
-
-	fn should_ignore_live_auxiliary_cursor_event_for_role(
-		mode: OverlayMode,
-		is_auxiliary_window: bool,
-	) -> bool {
-		matches!(mode, OverlayMode::Live) && is_auxiliary_window
-	}
-
-	fn current_device_cursor(&mut self) -> GlobalPoint {
-		self.sample_mouse_location()
-	}
-
-	fn current_frozen_interaction_cursor(&mut self) -> GlobalPoint {
-		if let Some((_, cursor)) = self.last_fresh_event_cursor() {
-			return cursor;
-		}
-		if let Some(cursor) = self.state.cursor {
-			return cursor;
-		}
-
-		let raw = self.current_device_cursor();
-
-		self.resolve_device_cursor_point(raw).map(|(_, cursor, _)| cursor).unwrap_or(raw)
-	}
-
-	fn trace_cursor_moved_with_mapping(&self, trace: CursorMoveTrace) {
-		if !tracing::enabled!(tracing::Level::TRACE) {
-			return;
-		}
-
-		let delta_x =
-			trace.global.x.abs_diff(trace.old_cursor.map_or(trace.global.x, |point| point.x));
-		let delta_y =
-			trace.global.y.abs_diff(trace.old_cursor.map_or(trace.global.y, |point| point.y));
-
-		tracing::trace!(
-			window_id = ?trace.window_id,
-			window_known = true,
-			window_position = ?trace.position,
-			old_cursor = ?trace.old_cursor,
-			device_cursor = ?trace.device_cursor,
-			event_cursor = ?trace.event_global,
-			source = trace.source.as_str(),
-			monitor_id = trace.monitor.id,
-			cursor_delta_x = delta_x,
-			cursor_delta_y = delta_y,
-			"CursorMoved coordinate source: {}.",
-			trace.source.as_str()
-		);
-	}
-
-	fn trace_frozen_selection_drag_cursor_move(
-		&self,
-		monitor: MonitorRect,
-		old_monitor: Option<MonitorRect>,
-		old_cursor: Option<GlobalPoint>,
-		timing: FrozenSelectionDragCursorMoveTiming,
-	) {
-		if !self.should_trace_frozen_selection_drag_timing() {
-			return;
-		}
-
-		tracing::trace!(
-			op = "overlay.frozen_selection_drag.cursor_move_timing",
-			monitor_id = monitor.id,
-			old_monitor_id = old_monitor.map(|target| target.id),
-			old_cursor = ?old_cursor,
-			cursor = ?self.state.cursor,
-			interaction = ?self.frozen_selection_drag.interaction,
-			frozen_rect_changed = timing.frozen_rect_changed,
-			cursor_update_us = timing.cursor_update_elapsed.as_micros(),
-			live_drag_update_us = timing.live_drag_update_elapsed.as_micros(),
-			frozen_drag_update_us = timing.frozen_drag_update_elapsed.as_micros(),
-			sync_cursor_icons_us = timing.sync_cursor_icons_elapsed.as_micros(),
-			request_samples_us = timing.request_samples_elapsed.as_micros(),
-			total_us = timing.total_elapsed.as_micros(),
-			overlay_window_count = self.windows.len(),
-			"Frozen selection drag cursor move timing."
-		);
-	}
-
-	fn should_trace_frozen_selection_drag_timing(&self) -> bool {
-		tracing::enabled!(tracing::Level::TRACE)
-			&& matches!(self.state.mode, OverlayMode::Frozen)
-			&& self.frozen_selection_drag.active
-	}
-
-	fn update_cursor_for_live_move(
-		&mut self,
-		old_monitor: Option<MonitorRect>,
-		old_cursor: Option<GlobalPoint>,
-		monitor: MonitorRect,
-		global: GlobalPoint,
-	) {
-		self.update_cursor_state(monitor, global);
-		self.update_hud_window_position(monitor, global);
-
-		if Self::live_hud_redraw_needed_for_cursor_update(old_cursor, global, old_monitor, monitor)
-		{
-			self.request_redraw_hud_window();
-		}
-		if self.should_try_pending_follow_window_move_on_live_cursor_update() {
-			self.maybe_apply_pending_hud_and_loupe_moves();
-		}
-		if matches!(self.state.mode, OverlayMode::Live) && self.use_fake_hud_blur() {
-			if self.state.live_bg_monitor != Some(monitor) {
-				self.state.live_bg_monitor = None;
-				self.state.live_bg_image = None;
-			}
-
-			self.maybe_request_live_bg(monitor);
-		}
-	}
-
-	fn request_cursor_move_samples(&mut self, monitor: MonitorRect, global: GlobalPoint) {
-		if !matches!(self.state.mode, OverlayMode::Live) {
-			return;
-		}
-		if self.pending_click_hit_test_request_id.is_some() {
-			return;
-		}
-
-		let is_dragging_window = matches!(self.state.mode, OverlayMode::Live)
-			&& self.left_mouse_button_down
-			&& self.left_mouse_button_down_monitor == Some(monitor);
-		let had_snapshot_update = if is_dragging_window || self.state.alt_held {
-			false
-		} else {
-			self.apply_live_hover_cache_state(monitor, global)
-		};
-		let sample_requested =
-			self.request_live_cursor_sample(monitor, global, self.state.alt_held);
-
-		if !is_dragging_window && !self.state.alt_held {
-			let _ = self.request_live_window_list_refresh_if_needed();
-		}
-
-		let apply = self.live_sample_request_redraw_intent(
-			had_snapshot_update,
-			sample_requested,
-			self.state.alt_held || self.loupe_window_visible,
-		);
-
-		if apply.any_changed() {
-			self.request_redraw_live_sample_targets(monitor, apply);
-		}
-	}
-
-	fn handle_left_mouse_input(
-		&mut self,
-		window_id: WindowId,
-		state: ElementState,
-	) -> OverlayControl {
-		let monitor = self
-			.windows
-			.get(&window_id)
-			.map(|w| w.monitor)
-			.or_else(|| self.active_cursor_monitor())
-			.or(self.state.monitor);
-		let Some(monitor) = monitor else {
-			return OverlayControl::Continue;
-		};
-
-		if matches!(self.state.mode, OverlayMode::Frozen) {
-			return self.handle_frozen_left_mouse_input(monitor, state);
-		}
-		if !matches!(self.state.mode, OverlayMode::Live) {
-			return OverlayControl::Continue;
-		}
-
-		match state {
-			ElementState::Pressed => {
-				if self.left_mouse_button_down {
-					return OverlayControl::Continue;
-				}
-
-				let raw_cursor = self.current_device_cursor();
-				let Some((press_monitor, press_global, _)) =
-					self.resolve_live_cursor_point(raw_cursor)
-				else {
-					self.left_mouse_button_down = true;
-					self.left_mouse_button_down_monitor = Some(monitor);
-					self.left_mouse_button_down_global = Some(raw_cursor);
-					self.state.drag_rect = None;
-					self.state.hovered_window_rect = None;
-
-					self.reset_toolbar_pointer_state();
-					self.request_redraw_for_monitor(monitor);
-
-					return OverlayControl::Continue;
-				};
-
-				self.left_mouse_button_down = true;
-				self.left_mouse_button_down_monitor = Some(press_monitor);
-				self.left_mouse_button_down_global = Some(press_global);
-				self.state.drag_rect = None;
-				self.state.hovered_window_rect = None;
-
-				self.reset_toolbar_pointer_state();
-				self.update_cursor_state(press_monitor, press_global);
-				self.update_hud_window_position(press_monitor, press_global);
-				self.request_redraw_for_monitor(press_monitor);
-
-				OverlayControl::Continue
-			},
-			ElementState::Released => {
-				let Some(start_monitor) = self.left_mouse_button_down_monitor else {
-					return OverlayControl::Continue;
-				};
-				let Some(start_global) = self.left_mouse_button_down_global else {
-					self.left_mouse_button_down = false;
-					self.left_mouse_button_down_monitor = None;
-
-					return OverlayControl::Continue;
-				};
-				let raw_cursor = self.current_device_cursor();
-				let (release_monitor, release_global) =
-					if let Some((release_monitor, release_global, _)) =
-						self.resolve_live_cursor_point(raw_cursor)
-					{
-						(release_monitor, release_global)
-					} else {
-						(start_monitor, start_global)
-					};
-
-				self.left_mouse_button_down = false;
-				self.left_mouse_button_down_monitor = None;
-				self.left_mouse_button_down_global = None;
-
-				let drag_rect = if start_monitor == release_monitor {
-					self.state.drag_rect.take()
-				} else {
-					None
-				};
-
-				if let Some(rect) = drag_rect
-					&& start_monitor == release_monitor
-					&& rect.monitor_id == release_monitor.id
-					&& rect.rect.width as f32 >= LIVE_DRAG_START_THRESHOLD_PX
-					&& rect.rect.height as f32 >= LIVE_DRAG_START_THRESHOLD_PX
-				{
-					self.begin_frozen_capture_with_rect(
-						release_monitor,
-						Some(rect.rect),
-						None,
-						Some(release_global),
-					);
-
-					return OverlayControl::Continue;
-				}
-
-				self.state.drag_rect = None;
-
-				self.request_click_capture_hit_test(release_monitor, release_global);
-
-				OverlayControl::Continue
-			},
-		}
-	}
-
-	fn handle_frozen_left_mouse_input(
-		&mut self,
-		monitor: MonitorRect,
-		state: ElementState,
-	) -> OverlayControl {
-		self.reset_toolbar_pointer_state();
-
-		if self.frozen_text_tool_active() {
-			match state {
-				ElementState::Pressed => {
-					let cursor = self.current_device_cursor();
-					let started_drag = self.begin_frozen_text_edit_drag_at(monitor, cursor);
-
-					if !started_drag {
-						let started = self.begin_frozen_text_edit_at(monitor, cursor);
-
-						if !started {
-							let _ = self.finish_frozen_text_editing(true);
-						}
-					}
-
-					self.sync_overlay_cursor_icons();
-				},
-				ElementState::Released => {
-					let stopped_drag = self.stop_frozen_text_edit_drag();
-
-					if stopped_drag {
-						self.sync_overlay_cursor_icons();
-					}
-				},
-			}
-
-			self.request_redraw_for_monitor(monitor);
-
-			return OverlayControl::Continue;
-		}
-		if self.frozen_text_edit.is_some() {
-			let _ = self.finish_frozen_text_editing(true);
-		}
-
-		match state {
-			ElementState::Pressed => {
-				let cursor = self.current_frozen_interaction_cursor();
-
-				if self.toolbar_state.selected_tool == FrozenToolbarTool::Pen {
-					let _ = self.begin_frozen_brush_stroke(cursor);
-				} else if !self.begin_frozen_selection_drag(cursor) {
-					let _ = self.begin_frozen_mosaic_drag(cursor);
-				}
-
-				self.sync_overlay_cursor_icons();
-			},
-			ElementState::Released => {
-				let _ = self.finish_frozen_brush_stroke();
-
-				self.stop_frozen_selection_drag();
-				self.stop_frozen_mosaic_drag();
-				self.sync_overlay_cursor_icons();
-			},
-		}
-
-		self.request_redraw_for_monitor(monitor);
-
-		OverlayControl::Continue
-	}
-
-	fn handle_ime_event(&mut self, window_id: WindowId, event: &Ime) -> OverlayControl {
-		if !matches!(self.state.mode, OverlayMode::Frozen) || self.frozen_text_edit.is_none() {
-			return OverlayControl::Continue;
-		}
-
-		let Some(monitor) =
-			self.windows.get(&window_id).map(|window| window.monitor).or(self.state.monitor)
-		else {
-			return OverlayControl::Continue;
-		};
-		let changed = self.apply_frozen_text_ime_event(event);
-
-		if changed {
-			self.sync_frozen_text_ime_cursor_area(monitor);
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		OverlayControl::Continue
-	}
-
-	fn apply_frozen_text_ime_event(&mut self, event: &Ime) -> bool {
-		match event {
-			Ime::Commit(text) => {
-				let generation = self.note_frozen_text_input_event();
-
-				self.append_text_to_frozen_edit_for_input_event(
-					FrozenTextInputSource::Ime,
-					generation,
-					text,
-				)
-			},
-			Ime::Preedit(text, cursor_range) => {
-				self.set_frozen_text_ime_preedit(Some(text.clone()), *cursor_range)
-			},
-			Ime::Disabled => self.set_frozen_text_ime_preedit(None, None),
-			Ime::Enabled => false,
-		}
-	}
-
-	fn handle_scroll_mouse_wheel(
-		&mut self,
-		window_id: WindowId,
-		delta: &MouseScrollDelta,
-	) -> OverlayControl {
-		if !self.scroll_capture.active || self.scroll_capture.paused {
-			return OverlayControl::Continue;
-		}
-
-		let Some(overlay_monitor) = self.windows.get(&window_id).map(|window| window.monitor)
-		else {
-			return OverlayControl::Continue;
-		};
-		let Some(scroll_monitor) = self.scroll_capture.monitor else {
-			return OverlayControl::Continue;
-		};
-		let Some(capture_rect) = self.scroll_capture.capture_rect_pixels else {
-			return OverlayControl::Continue;
-		};
-
-		if overlay_monitor != scroll_monitor {
-			return OverlayControl::Continue;
-		}
-
-		let cursor = self.current_device_cursor();
-		let cursor_pixels = scroll_monitor.local_u32_pixels(cursor);
-		let Some(cursor_pixels) = cursor_pixels else {
-			return OverlayControl::Continue;
-		};
-
-		if !capture_rect.contains(cursor_pixels) {
-			return OverlayControl::Continue;
-		}
-
-		self.record_scroll_capture_input_direction_from_overlay_wheel_at(delta, Instant::now());
-
-		#[cfg(target_os = "macos")]
-		{
-			let target_point = cursor;
-			let now = Instant::now();
-
-			self.arm_scroll_overlay_mouse_passthrough_window(now, "overlay_mouse_wheel");
-
-			let forwarded = self.forward_macos_scroll_wheel_event(
-				scroll_monitor,
-				cursor,
-				Some(cursor_pixels),
-				capture_rect,
-				target_point,
-				delta,
-			);
-
-			if !forwarded {
-				self.disarm_scroll_overlay_mouse_passthrough(now, "wheel_forward_failed");
-			}
-		}
-
-		OverlayControl::Continue
-	}
-
-	#[cfg(target_os = "macos")]
-	fn forward_macos_scroll_wheel_event(
-		&mut self,
-		scroll_monitor: MonitorRect,
-		cursor: GlobalPoint,
-		cursor_pixels: Option<(u32, u32)>,
-		capture_rect: RectPoints,
-		target_point: GlobalPoint,
-		delta: &MouseScrollDelta,
-	) -> bool {
-		let normalized = Self::normalize_macos_scroll_wheel_delta(
-			delta,
-			&mut self.scroll_capture.pixel_delta_residual,
-		);
-
-		if normalized.posted_x == 0 && normalized.posted_y == 0 {
-			return false;
-		}
-
-		if let Err(err) = macos_post_scroll_wheel_event(normalized, target_point) {
-			tracing::warn!(
-				op = "scroll_capture.wheel_forward_failed",
-				monitor_id = scroll_monitor.id,
-				cursor = ?cursor,
-				cursor_pixels = ?cursor_pixels,
-				capture_rect = ?capture_rect,
-				target_point = ?target_point,
-				raw_delta = ?delta,
-				normalized_delta_x = normalized.normalized_x,
-				normalized_delta_y = normalized.normalized_y,
-				posted_delta_x = normalized.posted_x,
-				posted_delta_y = normalized.posted_y,
-				pixel_residual_x = normalized.residual.x,
-				pixel_residual_y = normalized.residual.y,
-				error = %format!("{err:#}"),
-				"Failed to forward scroll wheel event."
-			);
-
-			self.state.set_error(format!("{err:#}"));
-			self.request_redraw_all();
-
-			return false;
-		}
-
-		tracing::info!(
-			op = "scroll_capture.wheel_forwarded",
-			monitor_id = scroll_monitor.id,
-			cursor = ?cursor,
-			cursor_pixels = ?cursor_pixels,
-			capture_rect = ?capture_rect,
-			target_point = ?target_point,
-			raw_delta = ?delta,
-			normalized_delta_x = normalized.normalized_x,
-			normalized_delta_y = normalized.normalized_y,
-			posted_delta_x = normalized.posted_x,
-			posted_delta_y = normalized.posted_y,
-			pixel_residual_x = normalized.residual.x,
-			pixel_residual_y = normalized.residual.y,
-			source_state_id = macos_hid_event_source_state_id(),
-			"Forwarded scroll wheel event."
-		);
-
-		true
-	}
-
-	#[cfg(target_os = "macos")]
-	fn normalize_macos_scroll_wheel_delta(
-		delta: &MouseScrollDelta,
-		residual: &mut MacOSScrollPixelResidual,
-	) -> MacOSScrollWheelEvent {
-		match delta {
-			MouseScrollDelta::LineDelta(x, y) => MacOSScrollWheelEvent {
-				units: KCG_SCROLL_EVENT_UNIT_LINE,
-				normalized_x: f64::from(*x),
-				normalized_y: f64::from(*y),
-				posted_x: x.round() as i32,
-				posted_y: y.round() as i32,
-				residual: *residual,
-			},
-			MouseScrollDelta::PixelDelta(delta) => {
-				let normalized_x = Self::normalize_macos_scroll_pixel_component(delta.x);
-				let normalized_y = Self::normalize_macos_scroll_pixel_component(delta.y);
-				let accumulated_x = residual.x + normalized_x;
-				let accumulated_y = residual.y + normalized_y;
-				let posted_x = accumulated_x.trunc() as i32;
-				let posted_y = accumulated_y.trunc() as i32;
-
-				*residual = MacOSScrollPixelResidual {
-					x: accumulated_x - f64::from(posted_x),
-					y: accumulated_y - f64::from(posted_y),
-				};
-
-				MacOSScrollWheelEvent {
-					units: KCG_SCROLL_EVENT_UNIT_PIXEL,
-					normalized_x,
-					normalized_y,
-					posted_x,
-					posted_y,
-					residual: *residual,
-				}
-			},
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	fn normalize_macos_scroll_pixel_component(value: f64) -> f64 {
-		if !value.is_finite() {
-			return 0.0;
-		}
-
-		let normalized = if value.abs() > MACOS_SCROLL_PIXEL_WRAP_THRESHOLD {
-			if value.is_sign_positive() {
-				value - MACOS_SCROLL_PIXEL_WRAP_MODULUS
-			} else {
-				value + MACOS_SCROLL_PIXEL_WRAP_MODULUS
-			}
-		} else {
-			value
-		};
-
-		normalized.clamp(-MACOS_SCROLL_PIXEL_DELTA_CLAMP, MACOS_SCROLL_PIXEL_DELTA_CLAMP)
-	}
-
-	fn scroll_capture_direction_from_wheel_delta(
-		delta: &MouseScrollDelta,
-	) -> Option<ScrollDirection> {
-		let vertical_delta = match delta {
-			MouseScrollDelta::LineDelta(_, y) => f64::from(*y),
-			MouseScrollDelta::PixelDelta(delta) => {
-				#[cfg(target_os = "macos")]
-				{
-					Self::normalize_macos_scroll_pixel_component(delta.y)
-				}
-				#[cfg(not(target_os = "macos"))]
-				{
-					delta.y
-				}
-			},
-		};
-
-		Self::scroll_capture_direction_from_delta_y(vertical_delta)
-	}
-
-	fn scroll_capture_direction_from_delta_y(vertical_delta: f64) -> Option<ScrollDirection> {
-		if vertical_delta < 0.0 {
-			Some(ScrollDirection::Down)
-		} else if vertical_delta > 0.0 {
-			Some(ScrollDirection::Up)
-		} else {
-			None
-		}
-	}
-
-	fn scroll_capture_direction_from_external_input_delta_y(
-		delta_y: f64,
-	) -> Option<ScrollDirection> {
-		if delta_y == 0.0 {
-			return None;
-		}
-
-		Self::scroll_capture_direction_from_delta_y(delta_y)
-	}
-
-	fn scroll_capture_motion_rows_from_wheel_delta(delta: &MouseScrollDelta) -> f64 {
-		match delta {
-			MouseScrollDelta::LineDelta(_, y) => f64::from(*y).abs(),
-			MouseScrollDelta::PixelDelta(delta) => {
-				#[cfg(target_os = "macos")]
-				{
-					Self::normalize_macos_scroll_pixel_component(delta.y).abs()
-				}
-				#[cfg(not(target_os = "macos"))]
-				{
-					delta.y.abs()
-				}
-			},
-		}
-	}
-
-	fn accumulate_scroll_capture_downward_motion_rows(&mut self, motion_rows: f64) {
-		if !motion_rows.is_finite() || motion_rows <= 0.0 {
-			return;
-		}
-
-		self.scroll_capture.downward_motion_rows_pending =
-			(self.scroll_capture.downward_motion_rows_pending + motion_rows.abs())
-				.clamp(0.0, SCROLL_CAPTURE_INPUT_MOTION_PRIOR_ROWS_MAX);
-	}
-
-	fn clear_scroll_capture_downward_motion_rows(&mut self) {
-		self.scroll_capture.downward_motion_rows_pending = 0.0;
-	}
-
-	fn consume_scroll_capture_downward_motion_rows(&mut self, consumed_rows: u32) {
-		if consumed_rows == 0 {
-			return;
-		}
-
-		let remaining = self.scroll_capture.downward_motion_rows_pending - f64::from(consumed_rows);
-
-		self.scroll_capture.downward_motion_rows_pending = remaining.max(0.0);
-	}
-
-	fn consume_scroll_capture_downward_motion_rows_for_outcome(
-		&mut self,
-		outcome: &ScrollObserveOutcome,
-	) {
-		if let ScrollObserveOutcome::Committed { direction: ScrollDirection::Down, growth_rows } =
-			outcome
-		{
-			self.consume_scroll_capture_downward_motion_rows(*growth_rows);
-		}
-	}
-
-	fn record_scroll_capture_input_direction_at(
-		&mut self,
-		direction: ScrollDirection,
-		gesture_active: bool,
-		at: Instant,
-	) {
-		self.scroll_capture.input_direction = Some(direction);
-		self.scroll_capture.input_direction_at = Some(at);
-		self.scroll_capture.input_gesture_active = gesture_active;
-
-		#[cfg(target_os = "macos")]
-		self.clear_incompatible_live_stream_stale_grace();
-	}
-
-	fn record_scroll_capture_input_direction_from_overlay_wheel_at(
-		&mut self,
-		delta: &MouseScrollDelta,
-		at: Instant,
-	) {
-		if let Some(direction) = Self::scroll_capture_direction_from_wheel_delta(delta) {
-			self.record_scroll_capture_input_direction_at(direction, false, at);
-
-			if matches!(direction, ScrollDirection::Down) {
-				self.accumulate_scroll_capture_downward_motion_rows(
-					Self::scroll_capture_motion_rows_from_wheel_delta(delta),
-				);
-			} else {
-				self.clear_scroll_capture_downward_motion_rows();
-			}
-		}
-	}
-
-	fn finish_scroll_capture_input_direction_at(&mut self, at: Instant) {
-		if self.scroll_capture.input_direction.is_some() {
-			self.scroll_capture.input_direction_at = Some(at);
-		} else {
-			self.scroll_capture.input_direction_at = None;
-		}
-
-		self.scroll_capture.input_gesture_active = false;
-
-		#[cfg(target_os = "macos")]
-		self.clear_incompatible_live_stream_stale_grace();
-	}
-
-	fn apply_scroll_capture_input_delta_y(
-		&mut self,
-		delta_y: f64,
-		gesture_active: bool,
-		gesture_ended: bool,
-		at: Instant,
-	) {
-		if let Some(direction) = Self::scroll_capture_direction_from_external_input_delta_y(delta_y)
-		{
-			if self.should_absorb_upward_external_input_into_active_downward_gesture(
-				direction,
-				gesture_active,
-			) {
-				self.record_scroll_capture_input_direction_at(
-					ScrollDirection::Down,
-					gesture_active,
-					at,
-				);
-				self.accumulate_scroll_capture_downward_motion_rows(delta_y.abs());
-			} else {
-				self.record_scroll_capture_input_direction_at(direction, gesture_active, at);
-
-				if matches!(direction, ScrollDirection::Down) {
-					self.accumulate_scroll_capture_downward_motion_rows(delta_y.abs());
-				} else {
-					self.clear_scroll_capture_downward_motion_rows();
-				}
-			}
-		}
-
-		if gesture_ended {
-			self.finish_scroll_capture_input_direction_at(at);
-		}
-	}
-
-	fn should_absorb_upward_external_input_into_active_downward_gesture(
-		&self,
-		direction: ScrollDirection,
-		gesture_active: bool,
-	) -> bool {
-		gesture_active
-			&& matches!(direction, ScrollDirection::Up)
-			&& self.scroll_capture.input_direction == Some(ScrollDirection::Down)
-			&& self.scroll_capture.downward_motion_rows_pending > 0.0
-	}
-
-	fn apply_external_scroll_input_delta_y(
-		&mut self,
-		global_x: f64,
-		global_y: f64,
-		delta_y: f64,
-		gesture_active: bool,
-		gesture_ended: bool,
-		at: Instant,
-	) {
-		if !self.scroll_capture.active || self.scroll_capture.paused {
-			return;
-		}
-
-		let Some(scroll_monitor) = self.scroll_capture.monitor else {
-			return;
-		};
-		let Some(capture_rect) = self.scroll_capture.capture_rect_pixels else {
-			return;
-		};
-		let cursor = GlobalPoint::new(global_x.round() as i32, global_y.round() as i32);
-		let Some(cursor_pixels) = scroll_monitor.local_u32_pixels(cursor) else {
-			return;
-		};
-
-		#[cfg(not(target_os = "macos"))]
-		if !capture_rect.contains(cursor_pixels) {
-			return;
-		}
-
-		#[cfg(target_os = "macos")]
-		let _cursor_inside_capture_rect = capture_rect.contains(cursor_pixels);
-
-		#[cfg(target_os = "macos")]
-		if delta_y != 0.0
-			&& !gesture_ended
-			&& !self.scroll_capture.overlay_mouse_passthrough_persistent
-		{
-			self.arm_scroll_overlay_mouse_passthrough_window(
-				Instant::now(),
-				"external_scroll_input",
-			);
-		}
-
-		self.apply_scroll_capture_input_delta_y(delta_y, gesture_active, gesture_ended, at);
-	}
-
-	fn scroll_capture_trace_snapshot_at(
-		&self,
-		observation_at: Instant,
-	) -> ScrollCaptureTraceSessionSnapshot {
-		ScrollCaptureTraceSessionSnapshot::capture(
-			self.scroll_capture.session.as_ref(),
-			self.scroll_capture_preview_dimensions(),
-			self.scroll_capture.input_direction,
-			self.scroll_capture.input_gesture_active,
-			self.scroll_capture.downward_motion_rows_pending,
-			self.scroll_capture_input_age_ms_at(observation_at),
-		)
-	}
-
-	#[cfg(test)]
-	fn scroll_capture_input_allows_observation(&self) -> bool {
-		self.scroll_capture_observation_block_reason().is_none()
-	}
-
-	#[cfg(test)]
-	fn scroll_capture_input_allows_growth(&self) -> bool {
-		self.scroll_capture_input_allows_observation()
-	}
-
-	#[cfg(test)]
-	fn scroll_capture_observation_block_reason(&self) -> Option<&'static str> {
-		self.scroll_capture_observation_block_reason_at(Instant::now())
-	}
-
-	fn scroll_capture_observation_block_reason_at(
-		&self,
-		observation_at: Instant,
-	) -> Option<&'static str> {
-		if self.scroll_capture.input_direction.is_none() {
-			return Some("missing_direction");
-		}
-		if self.scroll_capture.input_gesture_active {
-			return None;
-		}
-
-		let Some(input_direction_at) = self.scroll_capture.input_direction_at else {
-			return Some("missing_input_timestamp");
-		};
-
-		if observation_at.saturating_duration_since(input_direction_at)
-			> SCROLL_CAPTURE_INPUT_FRESHNESS
-		{
-			return Some("stale_input");
-		}
-
-		None
-	}
-
-	#[cfg(target_os = "macos")]
-	fn scroll_capture_input_age_ms(&self) -> Option<u64> {
-		self.scroll_capture_input_age_ms_at(Instant::now())
-	}
-
-	fn scroll_capture_input_age_ms_at(&self, observation_at: Instant) -> Option<u64> {
-		self.scroll_capture.input_direction_at.map(|input_direction_at| {
-			u64::try_from(observation_at.saturating_duration_since(input_direction_at).as_millis())
-				.unwrap_or(u64::MAX)
-		})
-	}
-
-	#[cfg(target_os = "macos")]
-	fn scroll_capture_should_force_stream_refresh_at(&self, now: Instant) -> bool {
-		if !self.scroll_capture_has_fresh_downward_backlog_at(now) {
-			return false;
-		}
-		if self.scroll_capture.input_gesture_active {
-			return false;
-		}
-
-		let Some(input_direction_at) = self.scroll_capture.input_direction_at else {
-			return false;
-		};
-
-		now.saturating_duration_since(input_direction_at) <= SCROLL_CAPTURE_INPUT_FRESHNESS
-	}
-
-	fn scroll_capture_has_fresh_downward_backlog_at(&self, now: Instant) -> bool {
-		if self.scroll_capture.input_direction != Some(ScrollDirection::Down)
-			|| self.scroll_capture.downward_motion_rows_pending <= 0.0
-		{
-			return false;
-		}
-
-		let Some(input_direction_at) = self.scroll_capture.input_direction_at else {
-			return false;
-		};
-
-		now.saturating_duration_since(input_direction_at) <= SCROLL_CAPTURE_INPUT_FRESHNESS
-	}
-
-	#[cfg(target_os = "macos")]
-	fn scroll_capture_should_schedule_stale_stream_refresh_at(&self, now: Instant) -> bool {
-		if !self.scroll_capture.input_gesture_active {
-			return true;
-		}
-
-		self.scroll_capture.last_stream_event_at.is_none_or(|last_stream_event_at| {
-			now.saturating_duration_since(last_stream_event_at)
-				>= SCROLL_CAPTURE_ACTIVE_GESTURE_STALE_REFRESH_DEAD_WINDOW
-		})
-	}
-
-	fn scroll_capture_should_allow_post_stall_burst_search_at(
-		&self,
-		frame_seq: u64,
-		now: Instant,
-	) -> bool {
-		self.scroll_capture.pending_post_stall_burst_after_seq.is_some_and(|after_seq| {
-			frame_seq > after_seq && self.scroll_capture_has_fresh_downward_backlog_at(now)
-		})
-	}
-
-	#[cfg(target_os = "macos")]
-	fn scroll_capture_should_arm_post_stall_burst_for_time_gap_at(
-		&self,
-		frame_captured_at: Instant,
-	) -> bool {
-		let Some(previous_captured_at) = self.scroll_capture.last_consumed_stream_frame_captured_at
-		else {
-			return false;
-		};
-
-		self.scroll_capture_has_fresh_downward_backlog_at(frame_captured_at)
-			&& frame_captured_at.saturating_duration_since(previous_captured_at)
-				>= SCROLL_CAPTURE_ACTIVE_GESTURE_STALE_REFRESH_DEAD_WINDOW
-	}
-
 	fn toolbar_pointer_state(
 		&mut self,
 		monitor: MonitorRect,
@@ -6369,244 +4706,6 @@ impl OverlaySession {
 			left_button_went_down,
 			left_button_went_up,
 		})
-	}
-
-	fn handle_frozen_text_key_event(&mut self, event: &KeyEvent) -> Option<OverlayControl> {
-		self.frozen_text_edit.as_ref()?;
-
-		if event.state != ElementState::Pressed {
-			return Some(OverlayControl::Continue);
-		}
-
-		let changed =
-			self.handle_frozen_text_pressed_key(&event.logical_key, event.text.as_deref());
-
-		if changed {
-			self.sync_text_input_ime_state();
-
-			if let Some(monitor) = self.state.monitor {
-				self.sync_frozen_text_ime_cursor_area(monitor);
-				self.request_redraw_for_monitor(monitor);
-			}
-		}
-
-		Some(OverlayControl::Continue)
-	}
-
-	fn handle_frozen_text_pressed_key(&mut self, logical_key: &Key, text: Option<&str>) -> bool {
-		match logical_key {
-			Key::Named(NamedKey::Escape) => {
-				let _ = self.finish_frozen_text_editing(false);
-
-				true
-			},
-			Key::Named(NamedKey::Enter) => {
-				if self.frozen_text_edit.as_ref().is_some_and(FrozenTextEditState::has_ime_preedit)
-				{
-					return false;
-				}
-				if self.keyboard_modifiers.shift_key() {
-					let generation = self.note_frozen_text_input_event();
-
-					self.append_text_to_frozen_edit_for_input_event(
-						FrozenTextInputSource::Key,
-						generation,
-						"\n",
-					)
-				} else {
-					let _ = self.finish_frozen_text_editing(true);
-
-					true
-				}
-			},
-			Key::Named(NamedKey::Backspace) => self.backspace_frozen_text_edit(),
-			_ if !self.keyboard_modifiers.control_key()
-				&& !self.keyboard_modifiers.super_key()
-				&& !self.keyboard_modifiers.alt_key() =>
-			{
-				let Some(text) = text else {
-					return false;
-				};
-				let generation = self.note_frozen_text_input_event();
-
-				self.append_text_to_frozen_edit_for_input_event(
-					FrozenTextInputSource::Key,
-					generation,
-					text,
-				)
-			},
-			_ => false,
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	/// Handles a host-level Escape hotkey press when the overlay is active.
-	pub fn handle_global_escape_hotkey(&mut self) -> OverlayControl {
-		if self.frozen_text_edit.is_some() {
-			let changed = self.handle_frozen_text_pressed_key(&Key::Named(NamedKey::Escape), None);
-
-			if changed {
-				self.sync_text_input_ime_state();
-
-				if let Some(monitor) = self.state.monitor {
-					self.sync_frozen_text_ime_cursor_area(monitor);
-					self.request_redraw_for_monitor(monitor);
-				}
-			}
-
-			return OverlayControl::Continue;
-		}
-		if self.scroll_capture.active {
-			return self.cancel_overlay("global_escape_hotkey_scroll_capture");
-		}
-
-		self.cancel_overlay("global_escape_hotkey")
-	}
-
-	fn handle_key_event(&mut self, event: &KeyEvent) -> OverlayControl {
-		if matches!(self.state.mode, OverlayMode::Frozen)
-			&& let Some(control) = self.handle_frozen_text_key_event(event)
-		{
-			return control;
-		}
-		if matches!(event.logical_key, Key::Named(NamedKey::Tab)) {
-			let pressed = event.state == ElementState::Pressed;
-
-			if self.apply_loupe_activation_key_event(pressed, event.repeat) {
-				return self.request_redraw_for_alt_state_change();
-			}
-
-			return OverlayControl::Continue;
-		}
-		if event.state != ElementState::Pressed {
-			return OverlayControl::Continue;
-		}
-		if event.repeat {
-			return OverlayControl::Continue;
-		}
-		if self.scroll_capture.active {
-			return self.handle_scroll_capture_key_event(event);
-		}
-
-		match &event.logical_key {
-			Key::Named(NamedKey::Escape) => self.cancel_overlay("escape_key"),
-			Key::Character(key_text)
-				if (key_text == "h" || key_text == "H")
-					&& self.plain_character_shortcut_available() =>
-			{
-				self.toolbar_state.visible = !self.toolbar_state.visible;
-
-				#[cfg(target_os = "macos")]
-				if self.toolbar_state.visible {
-					self.request_aux_window_creation_if_needed();
-				}
-
-				self.request_redraw_all();
-
-				OverlayControl::Continue
-			},
-			Key::Character(key_text)
-				if key_text.as_str().eq_ignore_ascii_case("c")
-					&& self.plain_character_shortcut_available() =>
-			{
-				self.auto_center_frozen_capture_rect();
-
-				OverlayControl::Continue
-			},
-			Key::Character(key_text)
-				if key_text.as_str().eq_ignore_ascii_case("s")
-					&& self.is_save_shortcut_pressed() =>
-			{
-				self.begin_png_action(PngAction::Save);
-
-				OverlayControl::Continue
-			},
-			Key::Character(key_text)
-				if key_text.as_str().eq_ignore_ascii_case("s")
-					&& self.plain_character_shortcut_available() =>
-			{
-				let available = self.scroll_capture_is_available();
-				let selection_ready = self.scroll_capture_selection_is_ready();
-
-				tracing::info!(
-				op = "scroll_capture.frozen_s_pressed",
-				available,
-				scroll_capture_active = self.scroll_capture.active,
-				selection_ready,
-				frozen_capture_source = ?self.frozen_capture_source,
-				state_mode = ?self.state.mode,
-				"Received `s` while frozen."
-				);
-
-				if selection_ready {
-					return self.start_scroll_capture();
-				}
-
-				OverlayControl::Continue
-			},
-			Key::Named(NamedKey::Space) => {
-				self.begin_png_action(PngAction::Copy);
-
-				OverlayControl::Continue
-			},
-			_ => OverlayControl::Continue,
-		}
-	}
-
-	fn is_save_shortcut_pressed(&self) -> bool {
-		#[cfg(target_os = "macos")]
-		{
-			self.keyboard_modifiers.super_key()
-		}
-		#[cfg(not(target_os = "macos"))]
-		{
-			self.keyboard_modifiers.control_key()
-		}
-	}
-
-	fn loupe_activation_shortcut_available(&self) -> bool {
-		matches!(self.state.mode, OverlayMode::Live)
-			&& !self.keyboard_modifiers.shift_key()
-			&& !self.keyboard_modifiers.alt_key()
-			&& !self.keyboard_modifiers.control_key()
-			&& !self.keyboard_modifiers.super_key()
-	}
-
-	fn plain_character_shortcut_available(&self) -> bool {
-		!self.loupe_activation_key_down
-			&& !self.keyboard_modifiers.alt_key()
-			&& !self.keyboard_modifiers.control_key()
-			&& !self.keyboard_modifiers.super_key()
-	}
-
-	fn handle_scroll_capture_key_event(&mut self, event: &KeyEvent) -> OverlayControl {
-		match &event.logical_key {
-			Key::Named(NamedKey::Escape) => self.cancel_overlay("scroll_capture_escape_key"),
-			Key::Named(NamedKey::Space) => {
-				self.begin_png_action(PngAction::Copy);
-
-				OverlayControl::Continue
-			},
-			Key::Character(key_text)
-				if key_text.as_str().eq_ignore_ascii_case("s")
-					&& self.is_save_shortcut_pressed() =>
-			{
-				self.begin_png_action(PngAction::Save);
-
-				OverlayControl::Continue
-			},
-			Key::Character(key_text) if key_text.as_str().eq_ignore_ascii_case("u") => {
-				self.undo_scroll_capture_append();
-
-				OverlayControl::Continue
-			},
-			Key::Character(key_text) if key_text.as_str().eq_ignore_ascii_case("p") => {
-				self.toggle_scroll_capture_paused();
-
-				OverlayControl::Continue
-			},
-			_ => OverlayControl::Continue,
-		}
 	}
 
 	fn current_export_image(&self) -> Option<RgbaImage> {
@@ -6711,212 +4810,6 @@ impl OverlaySession {
 		Some(Some(RectPoints::new(x, y, width, height)))
 	}
 
-	fn scroll_capture_selection_is_ready(&self) -> bool {
-		matches!(self.state.mode, OverlayMode::Frozen)
-			&& self.state.monitor.is_some()
-			&& self.state.frozen_capture_rect.is_some()
-			&& self.frozen_capture_source == FrozenCaptureSource::DragRegion
-			&& self.frozen_final_capture_ready()
-	}
-
-	fn scroll_capture_is_available(&mut self) -> bool {
-		if !self.scroll_capture_selection_is_ready() {
-			return false;
-		}
-
-		#[cfg(target_os = "macos")]
-		{
-			true
-		}
-		#[cfg(not(target_os = "macos"))]
-		{
-			false
-		}
-	}
-
-	fn toolbar_scroll_capture_slot_available(&self) -> bool {
-		if self.scroll_capture.active {
-			return true;
-		}
-
-		#[cfg(target_os = "macos")]
-		{
-			matches!(self.state.mode, OverlayMode::Frozen)
-				&& self.state.monitor.is_some()
-				&& self.state.frozen_capture_rect.is_some()
-				&& self.frozen_capture_source == FrozenCaptureSource::DragRegion
-		}
-
-		#[cfg(not(target_os = "macos"))]
-		{
-			false
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	fn try_prepare_scroll_capture_start(
-		&mut self,
-	) -> Option<(MonitorRect, RectPoints, RectPoints, RgbaImage)> {
-		if !self.scroll_capture_selection_is_ready() {
-			tracing::info!(
-				op = "scroll_capture.start_rejected",
-				reason = "selection_not_ready",
-				frozen_capture_source = ?self.frozen_capture_source,
-				state_mode = ?self.state.mode,
-				"Skipped starting scroll capture because the current frozen selection was not eligible."
-			);
-
-			self.state
-				.set_error(String::from("Scroll capture requires a dragged region selection."));
-			self.request_redraw_all();
-
-			return None;
-		}
-
-		let Some(monitor) = self.state.monitor else {
-			tracing::info!(
-				op = "scroll_capture.start_rejected",
-				reason = "missing_monitor",
-				"Skipped starting scroll capture because the frozen monitor was unavailable."
-			);
-
-			return None;
-		};
-		let Some(capture_rect_points) = self.state.frozen_capture_rect else {
-			tracing::info!(
-				op = "scroll_capture.start_rejected",
-				reason = "missing_capture_rect",
-				monitor_id = monitor.id,
-				"Skipped starting scroll capture because the frozen capture rect was unavailable."
-			);
-
-			return None;
-		};
-		let capture_rect_pixels = monitor.local_rect_to_pixels(capture_rect_points);
-		let Some(base_frame) =
-			self.cropped_monitor_frozen_region_image(monitor, capture_rect_pixels)
-		else {
-			tracing::info!(
-				op = "scroll_capture.start_rejected",
-				reason = "base_frame_unavailable",
-				monitor_id = monitor.id,
-				capture_rect_points = ?capture_rect_points,
-				capture_rect_pixels = ?capture_rect_pixels,
-				"Skipped starting scroll capture because the selected frozen region could not be read."
-			);
-
-			self.state
-				.set_error(String::from("Scroll capture could not read the selected region."));
-			self.request_redraw_all();
-
-			return None;
-		};
-
-		Some((monitor, capture_rect_points, capture_rect_pixels, base_frame))
-	}
-
-	#[cfg(target_os = "macos")]
-	fn build_scroll_capture_state(
-		&self,
-		monitor: MonitorRect,
-		capture_rect_points: RectPoints,
-		capture_rect_pixels: RectPoints,
-		base_frame: RgbaImage,
-	) -> Result<ScrollCaptureState> {
-		let use_worker_sampling = self.should_use_scroll_capture_worker_sampling();
-		let trace_recorder = ScrollCaptureTraceRecorder::from_env(
-			monitor,
-			capture_rect_pixels,
-			SCROLL_CAPTURE_PREVIEW_WIDTH_PX,
-			&base_frame,
-		);
-		let preview_latest_frame = Some(base_frame.clone());
-		let session = ScrollSession::new(base_frame, SCROLL_CAPTURE_PREVIEW_WIDTH_PX)?;
-		let preview_committed_image = Some(session.preview_image().clone());
-		let preview_display_image = preview_committed_image.clone();
-
-		Ok(ScrollCaptureState {
-			active: true,
-			paused: false,
-			monitor: Some(monitor),
-			#[cfg(target_os = "macos")]
-			capture_rect_points: Some(capture_rect_points),
-			capture_rect_pixels: Some(capture_rect_pixels),
-			input_direction: None,
-			input_direction_at: None,
-			input_gesture_active: false,
-			downward_motion_rows_pending: 0.0,
-			#[cfg(target_os = "macos")]
-			overlay_mouse_passthrough_active: false,
-			#[cfg(target_os = "macos")]
-			overlay_mouse_passthrough_persistent: false,
-			#[cfg(target_os = "macos")]
-			overlay_mouse_passthrough_until: None,
-			#[cfg(target_os = "macos")]
-			external_scroll_input_drain_reader: self
-				.scroll_capture
-				.external_scroll_input_drain_reader
-				.clone(),
-			last_external_scroll_input_seq: 0,
-			#[cfg(target_os = "macos")]
-			pixel_delta_residual: MacOSScrollPixelResidual::default(),
-			#[cfg(target_os = "macos")]
-			live_stream: (!use_worker_sampling).then(|| {
-				MacLiveFrameStream::with_scroll_capture_region_and_waker(
-					self.config.self_capture_exception_window_ids.clone(),
-					capture_rect_points,
-					capture_rect_pixels,
-					self.scroll_frame_waker.clone(),
-				)
-			}),
-			#[cfg(target_os = "macos")]
-			live_stream_backlog: VecDeque::new(),
-			last_stream_frame_seq: 0,
-			#[cfg(target_os = "macos")]
-			last_stream_frame_fingerprint: None,
-			#[cfg(target_os = "macos")]
-			consecutive_identical_stream_frames: 0,
-			#[cfg(target_os = "macos")]
-			last_consumed_stream_frame_captured_at: None,
-			#[cfg(target_os = "macos")]
-			last_stream_event_at: None,
-			#[cfg(target_os = "macos")]
-			last_stream_poll_at: None,
-			#[cfg(target_os = "macos")]
-			last_duplicate_stream_refresh_at: None,
-			pending_post_stall_burst_after_seq: None,
-			#[cfg(target_os = "macos")]
-			live_stream_stale_grace: None,
-			next_sample_at: Some(Instant::now() + SCROLL_CAPTURE_SAMPLE_INTERVAL),
-			next_request_id: 0,
-			inflight_request_id: None,
-			#[cfg(target_os = "macos")]
-			inflight_request_observation: None,
-			#[cfg(all(test, target_os = "macos"))]
-			force_worker_sampling_in_tests: false,
-			session: Some(session),
-			preview_committed_image,
-			preview_latest_frame,
-			preview_display_image,
-			retained_overlay_preview_image: None,
-			retained_overlay_preview_motion_rows_hint: None,
-			last_overlay_preview_motion_rows_hint: None,
-			last_overlay_preview_provisional_motion_rows_hint: None,
-			last_overlay_preview_existing_candidate_height: None,
-			last_overlay_preview_existing_candidate_motion_rows_hint: None,
-			last_overlay_preview_ledger_candidate_height: None,
-			last_overlay_preview_ledger_candidate_motion_rows_hint: None,
-			last_overlay_preview_retained_candidate_height: None,
-			last_overlay_preview_retained_candidate_motion_rows_hint: None,
-			last_overlay_preview_retained_hint_matches_motion_rows: false,
-			last_overlay_preview_fresh_latest_frame_can_drive: false,
-			last_overlay_preview_strong_unresolved_registration: false,
-			last_overlay_preview_latest_frame_present: false,
-			last_overlay_preview_used_provisional: false,
-			trace_recorder,
-		})
-	}
-
 	fn sync_frozen_toolbar_state(&mut self) {
 		self.toolbar_state.auto_center_available = self.frozen_auto_center_available();
 		self.toolbar_state.undo_available = self.frozen_undo_available();
@@ -6926,201 +4819,6 @@ impl OverlaySession {
 		// show the Scroll slot immediately, but keep it disabled until final_capture_ready flips.
 		self.toolbar_state.scroll_capture_available = self.toolbar_scroll_capture_slot_available();
 		self.toolbar_state.final_capture_ready = self.frozen_final_capture_ready();
-	}
-
-	fn start_scroll_capture(&mut self) -> OverlayControl {
-		if self.scroll_capture.active {
-			tracing::info!(
-				op = "scroll_capture.start_rejected",
-				reason = "already_active",
-				"Skipped starting scroll capture because a session is already active."
-			);
-
-			return OverlayControl::Continue;
-		}
-
-		#[cfg(not(target_os = "macos"))]
-		{
-			tracing::info!(
-				op = "scroll_capture.start_rejected",
-				reason = "unsupported_platform",
-				"Skipped starting scroll capture because the current platform is unsupported."
-			);
-
-			OverlayControl::Continue
-		}
-		#[cfg(target_os = "macos")]
-		{
-			let Some((monitor, capture_rect_points, capture_rect_pixels, base_frame)) =
-				self.try_prepare_scroll_capture_start()
-			else {
-				return OverlayControl::Continue;
-			};
-
-			if let Some(guard) = self.scroll_capture_start_guard.clone() {
-				match guard() {
-					Ok(true) => {},
-					Ok(false) => return OverlayControl::Continue,
-					Err(err) => {
-						self.state.set_error(format!("{err:#}"));
-						self.request_redraw_all();
-
-						return OverlayControl::Continue;
-					},
-				}
-			}
-			if let Some(hook) = self.scroll_capture_starting_hook.clone()
-				&& let Err(err) = hook()
-			{
-				self.state.set_error(format!("{err:#}"));
-				self.request_redraw_all();
-
-				return OverlayControl::Continue;
-			}
-
-			let base_frame_dimensions = base_frame.dimensions();
-
-			self.scroll_capture = match self.build_scroll_capture_state(
-				monitor,
-				capture_rect_points,
-				capture_rect_pixels,
-				base_frame,
-			) {
-				Ok(scroll_capture) => scroll_capture,
-				Err(err) => {
-					self.state.set_error(format!("{err:#}"));
-					self.request_redraw_all();
-
-					return OverlayControl::Continue;
-				},
-			};
-
-			if let Some(hook) = self.scroll_capture_started_hook.clone() {
-				hook();
-			}
-			if let Some(trace_recorder) = self.scroll_capture.trace_recorder.as_ref() {
-				tracing::info!(
-					op = "scroll_capture.trace_recording_enabled",
-					manifest_path = %trace_recorder.manifest_path().display(),
-					"Enabled scroll-capture live trace recording for this session."
-				);
-			}
-
-			tracing::info!(
-				op = "scroll_capture.start",
-				frozen_capture_source = ?self.frozen_capture_source,
-				monitor_id = monitor.id,
-				monitor_origin = ?monitor.origin,
-				monitor_size_points = ?(monitor.width, monitor.height),
-				monitor_scale_factor = monitor.scale_factor(),
-				capture_rect_points = ?capture_rect_points,
-				capture_rect_pixels = ?capture_rect_pixels,
-				base_frame_px = ?base_frame_dimensions,
-				"Entered scroll-capture mode."
-			);
-
-			self.request_aux_window_creation_if_needed();
-			self.sync_frozen_toolbar_state();
-			self.refresh_scroll_preview_committed_image();
-			self.refresh_scroll_preview_display_image();
-			self.sync_scroll_preview_segments();
-			self.position_scroll_preview_window(monitor);
-			self.update_scroll_toolbar_default_position(monitor);
-			self.set_scroll_overlay_mouse_passthrough_persistent(true, "scroll_capture_started");
-			self.focus_scroll_keyboard_window();
-			self.maybe_apply_pending_startup_aux_live_stream_filter_upgrade(monitor);
-
-			if let Some(preview) = self.scroll_preview_window.as_ref() {
-				preview.window.set_visible(true);
-				preview.window.request_redraw();
-			}
-			if let (Some(monitor), Some(live_stream)) =
-				(self.scroll_capture.monitor, self.scroll_capture.live_stream.as_ref())
-			{
-				live_stream.prime_monitor_nonblocking(monitor);
-			}
-
-			self.request_redraw_for_monitor(monitor);
-
-			OverlayControl::Continue
-		}
-	}
-
-	fn toggle_scroll_capture_paused(&mut self) {
-		if !self.scroll_capture.active {
-			return;
-		}
-
-		self.scroll_capture.paused = !self.scroll_capture.paused;
-
-		#[cfg(target_os = "macos")]
-		if self.scroll_capture.paused {
-			self.set_scroll_overlay_mouse_passthrough_persistent(false, "paused");
-		}
-		if !self.scroll_capture.paused {
-			#[cfg(target_os = "macos")]
-			{
-				self.set_scroll_overlay_mouse_passthrough_persistent(true, "resumed");
-
-				if let (Some(monitor), Some(live_stream)) =
-					(self.scroll_capture.monitor, self.scroll_capture.live_stream.as_ref())
-				{
-					live_stream.prime_monitor_nonblocking(monitor);
-				}
-			}
-			#[cfg(not(target_os = "macos"))]
-			{
-				self.scroll_capture.next_sample_at =
-					Some(Instant::now() + SCROLL_CAPTURE_SAMPLE_INTERVAL);
-			}
-		}
-
-		self.request_redraw_scroll_preview_window();
-	}
-
-	fn prepare_active_scroll_capture_output(&mut self) {
-		if !self.scroll_capture.active {
-			return;
-		}
-
-		self.maybe_tick_scroll_capture();
-		self.refresh_scroll_preview_committed_image();
-		self.refresh_scroll_preview_display_image();
-		self.sync_scroll_preview_segments();
-	}
-
-	fn undo_scroll_capture_append(&mut self) {
-		if !self.scroll_capture.active {
-			return;
-		}
-
-		let Some(session) = self.scroll_capture.session.as_mut() else {
-			return;
-		};
-
-		if !session.undo_last_append() {
-			return;
-		}
-
-		self.refresh_scroll_preview_committed_image();
-		self.clear_scroll_capture_inflight_request();
-
-		#[cfg(target_os = "macos")]
-		{
-			if let (Some(monitor), Some(live_stream)) =
-				(self.scroll_capture.monitor, self.scroll_capture.live_stream.as_ref())
-			{
-				live_stream.prime_monitor_nonblocking(monitor);
-			}
-		}
-		#[cfg(not(target_os = "macos"))]
-		{
-			self.scroll_capture.next_sample_at =
-				Some(Instant::now() + SCROLL_CAPTURE_SAMPLE_INTERVAL);
-		}
-
-		self.refresh_scroll_preview_display_image();
-		self.sync_scroll_preview_segments();
 	}
 
 	fn begin_png_action(&mut self, action: PngAction) {
@@ -7230,111 +4928,6 @@ impl OverlaySession {
 		}
 
 		self.handle_overlay_window_redraw(window_id)
-	}
-
-	#[cfg(target_os = "macos")]
-	#[cfg(target_os = "macos")]
-	fn set_scroll_overlay_mouse_passthrough(&self, passthrough: bool) {
-		for overlay_window in self.windows.values() {
-			let _ = overlay_window.window.set_cursor_hittest(!passthrough);
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	fn set_scroll_overlay_mouse_passthrough_state(
-		&mut self,
-		now: Instant,
-		passthrough: bool,
-		reason: &'static str,
-	) {
-		if self.scroll_capture.overlay_mouse_passthrough_active == passthrough {
-			return;
-		}
-
-		self.set_scroll_overlay_mouse_passthrough(passthrough);
-
-		self.scroll_capture.overlay_mouse_passthrough_active = passthrough;
-
-		tracing::info!(
-			op = if passthrough {
-				"scroll_capture.mouse_passthrough_armed"
-			} else {
-				"scroll_capture.mouse_passthrough_disarmed"
-			},
-			reason,
-			passthrough,
-			deadline_in_ms = self.scroll_capture.overlay_mouse_passthrough_until.map(|deadline| {
-				u64::try_from(deadline.saturating_duration_since(now).as_millis())
-					.unwrap_or(u64::MAX)
-			}),
-			"Updated scroll-capture mouse passthrough state."
-		);
-	}
-
-	#[cfg(target_os = "macos")]
-	fn set_scroll_overlay_mouse_passthrough_persistent(
-		&mut self,
-		passthrough: bool,
-		reason: &'static str,
-	) {
-		let now = Instant::now();
-
-		self.scroll_capture.overlay_mouse_passthrough_persistent = passthrough;
-		self.scroll_capture.overlay_mouse_passthrough_until = None;
-
-		self.set_scroll_overlay_mouse_passthrough_state(now, passthrough, reason);
-	}
-
-	#[cfg(target_os = "macos")]
-	fn arm_scroll_overlay_mouse_passthrough_window(&mut self, now: Instant, reason: &'static str) {
-		if self.scroll_capture.overlay_mouse_passthrough_persistent {
-			return;
-		}
-
-		let deadline = now + SCROLL_CAPTURE_MOUSE_PASSTHROUGH_IDLE_GRACE;
-		let was_active = self.scroll_capture.overlay_mouse_passthrough_active;
-
-		self.scroll_capture.overlay_mouse_passthrough_until = Some(deadline);
-
-		self.set_scroll_overlay_mouse_passthrough_state(now, true, reason);
-
-		if was_active {
-			tracing::info!(
-				op = "scroll_capture.mouse_passthrough_extended",
-				reason,
-				deadline_in_ms = u64::try_from(deadline.saturating_duration_since(now).as_millis())
-					.unwrap_or(u64::MAX),
-				"Extended scroll-capture mouse passthrough window."
-			);
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	fn disarm_scroll_overlay_mouse_passthrough(&mut self, now: Instant, reason: &'static str) {
-		self.scroll_capture.overlay_mouse_passthrough_persistent = false;
-		self.scroll_capture.overlay_mouse_passthrough_until = None;
-
-		self.set_scroll_overlay_mouse_passthrough_state(now, false, reason);
-	}
-
-	#[cfg(target_os = "macos")]
-	fn sync_scroll_overlay_mouse_passthrough_window(&mut self, now: Instant) {
-		if self.scroll_capture.overlay_mouse_passthrough_persistent {
-			return;
-		}
-		if !self.scroll_capture.overlay_mouse_passthrough_active {
-			return;
-		}
-
-		let Some(deadline) = self.scroll_capture.overlay_mouse_passthrough_until else {
-			self.set_scroll_overlay_mouse_passthrough_state(now, false, "missing_deadline");
-
-			return;
-		};
-
-		if deadline <= now {
-			self.disarm_scroll_overlay_mouse_passthrough(now, "idle_timeout");
-		}
 	}
 
 	#[cfg(target_os = "macos")]
@@ -7455,49 +5048,6 @@ impl OverlaySession {
 		);
 
 		macos_make_window_key(target_window);
-	}
-
-	#[cfg(target_os = "macos")]
-	fn focus_scroll_keyboard_window(&self) {
-		macos_activate_app();
-
-		let target_window = if let Some(toolbar_window) = self.toolbar_window.as_ref() {
-			Some(toolbar_window.window.as_ref())
-		} else if let Some(preview_window) = self.scroll_preview_window.as_ref() {
-			Some(preview_window.window.as_ref())
-		} else {
-			self.windows
-				.values()
-				.find(|overlay_window| Some(overlay_window.monitor) == self.scroll_capture.monitor)
-				.map(|overlay_window| overlay_window.window.as_ref())
-		};
-		let Some(target_window) = target_window else {
-			return;
-		};
-
-		macos_make_window_key(target_window);
-	}
-
-	fn update_scroll_toolbar_default_position(&mut self, monitor: MonitorRect) {
-		if !self.scroll_capture.active || self.toolbar_state.dragging {
-			return;
-		}
-
-		let screen_rect =
-			Rect::from_min_size(Pos2::ZERO, Vec2::new(monitor.width as f32, monitor.height as f32));
-		let preview_rect = self.scroll_preview_local_rect(monitor);
-		let toolbar_size = WindowRenderer::frozen_toolbar_size(&self.toolbar_state);
-		let toolbar_pos = WindowRenderer::frozen_toolbar_default_pos(
-			screen_rect,
-			preview_rect,
-			toolbar_size,
-			self.config.toolbar_placement,
-		);
-
-		self.toolbar_state.default_slot_position = Some(toolbar_pos);
-		self.toolbar_state.floating_position = Some(toolbar_pos);
-
-		let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
 	}
 
 	fn maybe_recenter_frozen_toolbar_default_slot(&mut self, monitor: MonitorRect) -> bool {
@@ -7918,31 +5468,6 @@ impl OverlaySession {
 			last_event_detail = ?self.event_loop_last_progress_detail,
 			"Beginning overlay exit cleanup."
 		);
-	}
-
-	fn finalize_scroll_capture_for_exit(&mut self) {
-		if self.scroll_capture.active {
-			self.maybe_tick_scroll_capture();
-			self.refresh_scroll_preview_committed_image();
-			self.refresh_scroll_preview_display_image();
-			self.sync_scroll_preview_segments();
-		}
-
-		let scroll_capture_final_snapshot = self.scroll_capture_trace_snapshot_at(Instant::now());
-		let final_preview_image = self.current_scroll_preview_render_image();
-
-		if let (Some(trace_recorder), Some(session)) =
-			(self.scroll_capture.trace_recorder.as_mut(), self.scroll_capture.session.as_ref())
-		{
-			let final_preview_image =
-				final_preview_image.unwrap_or_else(|| session.preview_image().clone());
-
-			trace_recorder.finalize_session(
-				session,
-				&final_preview_image,
-				scroll_capture_final_snapshot,
-			);
-		}
 	}
 
 	fn reset_runtime_for_exit(&mut self) {
